+++ CHANGELOG.md
@@ -5,6 +5,9 @@ Changelog
 ### Changed
 - `Field::getObjectLabel` is now public (was protected).
 
+### Added
+- EnvironmentalTableTrait to assist w/ environment specific tables 
+
 ## 2.2.1 - 2019-05-08
 ### Fixed
 - Migration would attempt to add 'name' column even if it already existed.